Update Flight_Chart_Program.pde
I added the crash factor function but all it does right now is display the explosion for when a plane crashes at random intervals since the crash factor is determined by a random chance. Plan is to have this run in the plane class so in the draw function we could have various functions like plane1.checkCrash();, plane2.checkCrash();, etc. I was also planning to have the vector of the plane's coordinate set back to the airport within the checkCrash(); function but I haven't been working on the plane class so I couldn't really make that change. Also for this part of the code:

  if(exwidth>=20){
    exwidth=0;
    crashFactor=0.01;
  }

will have to be altered to check the exwidth's of like the individual planes, we could probably make it a function in the plane class that could run in the draw function as well. Also the checkCrash(); functions should be ran last in the draw function to avoid interference with any other things in the draw function. I'll show you guys what I have tomorrow and we could adjust it accordingly.