Create crashFactor updated version
Here's the updated code I have for the crash factor. The draw function should have the checkExtracker(); and resetExTracker(); functions ran (ran lastly preferably) in it for each of the planes, since those functions will be defined in the plane class. You can see the general structure of the code shown so it should work if you include those 2 functions in the draw function. The exTracker variable should also be unique to each plane so that variable should be declared in the plane class, as a field or the planes. Hope that helps