Create airPlaneSimpMain with updated crash factor
I messed around with the code and got the crash factor working quite as planned. I had to change the code for the main and plane class to get it to work so I'll post the code for those (we can decide later I we want to use them). It doesn't work with the GUI but that's because I didn't set the values of the crash factor as the values of the slider. I was also thinking of changing the slider for the crash factor to a text entry box where the user can enter a number for the crash factor since I feel like expressing a number as low as 0.01 (which in this case is 1%) or even lower would be more difficult to represent on a slider but to be fair though I could be wrong about that. Also when I copied the code of the main and plane class it gave me an empty window for the GUI when I accessed it through the tools window, just thought I should let you know.