Create Plane class with updated crash factor
Anyways, this is the code for the updated plane class. Another thing I wanted to add that I did not say before is that hypothetically, the code should work as planned if you put the updated main and plane classes into your code. But! if somehow I forgot that I changed the other classes and you run it to find that it is missing something tell me and I'll post the rest of the code (assuming I changed something and forgot to post it).